requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Retrieval
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Evaluation
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0

# Utilities
pyyaml>=6.0
tiktoken>=0.5.0

